Complete the scaffolding: move aces/process_composite, eval_suite/eval.py, sample_codes; write top-level README
Per Matthew: treat #90 as the one-stop scaffolding PR so collaborators can see the whole
relocation in a single diff.

Additional moves landed here:
- aces_to_eq/ + process_composite/ → predict/external_tasks/ (both tools plus their configs
  under one 'predict a task outside EQ's vocabulary' umbrella per #62).  get_per_code_from_composite_config.yaml
  moves out of eval_suite/conf/ into predict/external_tasks/configs/ to colocate with its consumer.
- eval_suite/ + eval.py → evaluate/ (pure file move + Hydra CONFIGS path update; the #83
  consolidation is separate future work).  pyproject.toml entry points point at the new paths.
- sample_codes/ → paper_experiments/sample_codes/ (the dataset-agnostic rewrite stays in #85).

Top-level README rewritten from scratch to reflect the final layout — supersedes the interim
snapshot previously in PR #69.  Subsumes what that PR was doing so it can close without
merging.  Also adds README.md for predict/, predict/external_tasks/, and evaluate/ and
updates paper_experiments/README.md to mention the moved sample_codes/.

191 tests pass.  All entry points still resolve.  No behavior change.
