+++ src/every_query/paper_experiments/README.md
@@ -20,6 +20,13 @@ Anything on the normal-usage path: `preprocessing/`, `generate_tasks/`, `train/`
 
 Cluster-specific operator tooling (SLURM submit wrappers, GCS upload scripts) also doesn't live here — that category already left the repo in [#77](https://github.com/payalchandak/EveryQuery/pull/77).
 
+## What lives here today
+
+- **`sample_codes/`** — scripts that sample query codes from a MEDS dataset into `train_codes`
+    and `eval_codes` YAMLs, with an ID/OOD split axis. Moved here in #90 alongside the rest of
+    the submodule scaffolding; the pending dataset-agnostic rewrite (removing hardcoded MIMIC
+    paths, taking `metadata_dir` as a CLI arg) is tracked as [#85](https://github.com/payalchandak/EveryQuery/issues/85).
+
 ## Optional dependencies
 
 If a specific experiment needs heavyweight plotting deps (matplotlib, seaborn, plotly), those go under `[project.optional-dependencies].paper` in `pyproject.toml` rather than polluting the core install. Parked until we see the real dep footprint.
@@ -27,5 +34,5 @@ If a specific experiment needs heavyweight plotting deps (matplotlib, seaborn, p
 ## Related
 
 - Parent refactor umbrella: [#54](https://github.com/payalchandak/EveryQuery/issues/54)
-- Planned `sample_codes/` move into here: [#85](https://github.com/payalchandak/EveryQuery/issues/85)
+- `sample_codes/` dataset-agnostic rewrite: [#85](https://github.com/payalchandak/EveryQuery/issues/85)
 - Plan context for this split: [refactor plan v3](https://github.com/payalchandak/EveryQuery/issues/54#issuecomment-4271122440)
