+++ src/every_query/predict/external_tasks/aces_to_eq.py
@@ -56,10 +56,10 @@ def create_eq_task_df(
         logger.info(f"Wrote to {out_fp}")
 
 
-ACES_CONFIGS = str(files("every_query") / "aces_to_eq")
+ACES_CONFIGS = str(files("every_query") / "predict" / "external_tasks" / "configs")
 
 
-@hydra.main(version_base="1.3", config_path=ACES_CONFIGS, config_name="config.yaml")
+@hydra.main(version_base="1.3", config_path=ACES_CONFIGS, config_name="aces_to_eq.yaml")
 def main(cfg: DictConfig) -> None:
     all_eq_shards = sorted(
         Path(cfg.eq_tasks_all_dir).glob("*.parquet"),
